requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
google-genai
requests
python-dotenv